refactor(automations): deliverWebhook joins the server package with a required injected sender (ADR-063 P1)
The send-and-record wrapper moves to clients/http/deliver-webhook; the
sender default is gone — the settlement intent handler now threads the
app's appWebhookSender explicitly, and the graph-alert path already
injected its own. The unit suite moves with it.
